Remove code duplication between remoteReplicatedMap source and sink
